practice settings overhaul
